requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
langchain-groq==0.0.1
langchain>=0.1.0
langchain-core>=0.1.0
pydantic==2.6.1
logging>=0.5.1.2
python-json-logger>=2.0.7